+++ testcases/fail/explicit_if_else.sv
@@ -1,0 +1,4 @@
+module A;
+always_ff if (x) y <= 0;
+always_comb if (x) y = 0;
+endmodule