+++ testcases/fail/localparam_explicit_type.sv
@@ -1,0 +1,3 @@
+module A;
+localparam a = 0;
+endmodule