+++ testcases/fail/sequential_block_in_always_ff.sv
@@ -1,0 +1,25 @@
+module a;
+  always_ff @(posedge clk) begin
+    a <= z;
+  end
+
+  always_ff @(posedge clk)
+    if (bar) begin
+      b <= z;
+    end
+
+  always_ff @(posedge clk)
+    if (bar) c <= z;
+    else begin
+      c <= z;
+    end
+
+  always_ff @(posedge clk)
+    case (bar)
+      one: begin
+        d <= z;
+      end
+      two: d <= z;
+      default: d <= z;
+    endcase
+endmodule