+++ testcases/pass/explicit_case_default.sv
@@ -1,0 +1,14 @@
+module A;
+always_comb begin
+    case (x)
+        1: y = 0;
+        default: y = 0;
+    endcase
+end
+always_ff begin
+    case (x)
+        1: y = 0;
+        default: y = 0;
+    endcase
+end
+endmodule