+++ testcases/pass/explicit_if_else.sv
@@ -1,0 +1,8 @@
+module A;
+always_ff
+  if (x) y <= 0;
+  else   y <= z;
+always_comb
+  if (x) y = 0;
+  else   y = z;
+endmodule