+++ testcases/pass/sequential_block_in_always_latch.sv
@@ -1,0 +1,15 @@
+module a;
+  always_latch
+    if (foo) e <= z;
+
+  always_latch
+    if (foo) f <= z;
+    else     f <= z;
+
+  always_latch
+    case (foo)
+      one:     g <= z;
+      two:     g <= z;
+      default: g <= z;
+    endcase
+endmodule